Changed the order of filename/sequence for the input of qimcdespot/ qisignal so that they match. Updated scripts to reflect this, pulled SSFP tests out into a separate script.

+++ Test/test_2C.sh
@@ -1,19 +1,17 @@
 #!/bin/bash -e
 
 # Tobias Wood 2015
-# Simple test script for DESPOT programs
-
-# Tests whether programs run successfully on toy data
-
-# First, create input data
+# Simple test scripts for QUIT programs
 
 source ./test_common.sh
-SILENCE_TESTS="0"
+SILENCE_TESTS="1"
 
 DATADIR="2C"
 mkdir -p $DATADIR
 cd $DATADIR
-rm *
+if [ "$(ls -A ./)" ]; then
+    rm *
+fi
 
 DIMS="5 5 3"
 
@@ -36,7 +34,21 @@ SSFP_FLIP="12 16 21 27 33 40 51 68 "
 SSFP_PC="180 0"
 SSFP_TR="0.005"
 
-run_test "CREATE_SIGNALS" $QUITDIR/qisignal --2 -n -v << END_MCSIG
+INPUT="$SPGR_FILE
+SPGR
+$SPGR_FLIP
+$SPGR_TR
+$SSFP_FILE
+SSFP
+$SSFP_FLIP
+$SSFP_PC
+$SSFP_TR
+END
+"
+
+echo "$INPUT" > mcd_input.txt
+
+run_test "CREATE_SIGNALS" $QUITDIR/qisignal --2 -v << END_SIG
 PD.nii
 T1_m.nii
 T2_m.nii
@@ -46,33 +58,8 @@ tau_m.nii
 f_m.nii
 f0.nii
 B1.nii
-SPGR
-$SPGR_FLIP
-$SPGR_TR
-$SPGR_FILE
-SSFP
-$SSFP_FLIP
-$SSFP_PC
-$SSFP_TR
-$SSFP_FILE
-END
-END_MCSIG
-
-# Save these for mcd input to make debugging easy.
-# Currently the sequence type/filename are the opposite to qisignal due to history/idiocy
-# This needs to be changed for a later release
-ALL_PARS="SPGR
-$SPGR_FILE
-$SPGR_FLIP
-$SPGR_TR
-SSFP
-$SSFP_FILE
-$SSFP_FLIP
-$SSFP_PC
-$SSFP_TR
-END"
-
-echo "$ALL_PARS" > mcd_input.txt
+$INPUT
+END_SIG
 
 function run() {
 PREFIX="$1"
